+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   `sendsTo` configuration.
 * Graph builder returns error if nodes remain unconnected
 * Graph builder returns error if a node sends data to itself
+* A node config implementing `Enabler` interface can allow users disabling nodes without requiring pointers.
 
 ## Breaking changes
 * High-level graph API add contexts to the `graph.Run(context.Context)` library, and